requirements.txt
aiogram~=3.26.0
python-dotenv~=1.2.2
aiohttp~=3.13.3